// Bun-Servers/03-socket-maps/src/services/clients.service.ts
import type { ClientMovePayload, ClientRegisterPayload } from "../schemas/websocket-message.schema";
import { ClientsStore } from "../store/clients.store";
import type { ClientMarker } from "../types";


class ClientsService {
  private readonly clientsStore: ClientsStore;

  constructor() {
    this.clientsStore = new ClientsStore();
  }

  getAllClient() {
    return this.clientsStore.getAll();
  }

  registerClient(input: ClientRegisterPayload): { error: string } | ClientMarker {
    if (this.clientsStore.has(input.clientId)) {
        return { error: 'Client already register'}
    };

    const client: ClientMarker = {
        ...input,
        updateAt: Date.now(),
        color: input.color ||'gray'
    }

    this.clientsStore.add(client);
    return client;
  }

  clientMoved(clientId: string, input: ClientMovePayload): { error: string } | ClientMarker {
    const client = this.clientsStore.getById(clientId);

    if (!client) return { error: 'Client not register.'}    

    const updatedClient = this.clientsStore.updateCoords(clientId, input.coords);

    if ( !updatedClient ) return { error: 'Coldnot move client. Client not found' };
    return updatedClient;
  }

  removeClient(clientId: string) {
    return this.clientsStore.remove(clientId);
  }
}

export const clientsService = new ClientsService();